Update batch float

+++ bonito/crf/basecall.py
@@ -39,7 +39,7 @@ def compute_scores(model, batch, beam_width=32, beam_cut=100.0, scale=1.0, offse
     # dtype = torch.float16 if half_supported() else torch.float32
     # scores = model(batch.to(dtype).to(device))
     print("Sending Batch to Device", file=sys.stderr)
-    batch = batch.to(torch.bfloat16).to(device)
+    batch = batch.to(device)
     
     scores = model(batch)
     print("Sent batch to device and evaluated", file=sys.stderr)
